# api/swagger.yaml
openapi: 3.0.3
info:
  title: PetSitter API
  version: "0.1"
servers:
- url: /
...
paths:
  /users:
    post:
      tags:
      - users
      summary: Register User
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        "201":
          description: Created
          headers:
            Location:
              style: simple
              explode: false
              schema:
                type: string
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        password:
          type: string
        full_name:
          type: string
        roles:
          type: array
          items:
            type: string
      example:
        password: password
        full_name: full_name
        roles:
        - roles
        - roles
        id: 0
        email: email
...